[#1189] Move views classes back to simply views.rb
